requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==24.1.0
ale-py==0.8.1
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
apriltag==0.0.16
asttokens==3.0.0
astunparse==1.6.3
audioop-lts==0.2.1
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
blinker==1.9.0
box2d-py==2.3.5
Brotli==1.1.0
certifi==2024.12.14
-e git+https://github.com/bitcraze/crazyflie-clients-python@0bb1dc7d669ffeac1e0b568a46a785a687aae546#egg=cfclient
cflib==0.1.23
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
dash==3.0.3
decorator==5.1.1
executing==2.1.0
Farama-Notifications==0.0.4
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
Flask==3.0.3
flatbuffers==24.12.23
freetype-py==2.4.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.39.0
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.68.1
gtsam==4.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.16.0
h5py==3.12.1
hf-xet==1.1.5
hsluv==5.0.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib-resources==6.0.1
importlib_metadata==8.6.1
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyterlab_widgets==3.0.13
keras==3.7.0
keras-nightly==3.8.0.dev2025010303
libclang==18.1.1
libusb-package==1.0.26.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
namex==0.0.8
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.20.1
opencv-python==4.9.0.80
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pupil-apriltags==1.0.4.post10
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pydicom==3.0.1
pydub==0.25.1
pygame==2.5.2
Pygments==2.18.0
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pyqtgraph==0.13.3
pyquaternion==0.9.9
PySDL2==0.9.16
pysdl2-dll==2.24.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.32.3
retrying==1.3.4
rich==14.1.0
ruff==0.12.7
safehttpx==0.1.6
scikit-learn==1.6.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools @ file:///opt/homebrew/Cellar/python%403.11/3.11.13/Frameworks/Python.framework/Versions/3.11/lib/python3.11/ensurepip/_bundled/setuptools-80.9.0-py3-none-any.whl#sha256=f9be7553539415c7581772970da2d16361fb28786fd5221cfa8319bb44baecbc
shellingham==1.5.4
Shimmy==0.2.1
sip @ file:///private/tmp/sip-20231005-5173-81amdm/sip-6.7.11
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
swig==4.3.0
sympy==1.13.1
tb-nightly==2.19.0a20250102
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_nightly==2.19.0.dev20241216
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
vispy==0.13.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.6
wheel @ file:///opt/homebrew/Cellar/python%403.11/3.11.13/libexec/wheel-0.45.1-py3-none-any.whl#sha256=1861537b50a755fb89d55eee6ac81b43aa0120c2a941b039a1effa190720c86f
widgetsnbextension==4.0.13
wrapt==1.17.0
zipp==3.21.0
